requirement.txt
datetime
discord.py
requests
genshin.py
dotenv